Create new Box class
Box is the main container of this game
