+++ src/net/honux/neko/Coin.java
@@ -8,8 +8,8 @@ public class Coin extends GameObject{
 
     private final String NORMAL = "NORMAL";
     private boolean show = true;
-    public Coin(MainWindow w, int x, int y, int scale) {
-        super(w, w.DELAY, scale);
+    public Coin(Box box, int x, int y, int scale) {
+        super(box, box.DELAY, scale);
         status=NORMAL;
         setPosition(x, y);
     }
@@ -25,7 +25,7 @@ public class Coin extends GameObject{
 
     @Override
     public void update(int frame) {
-        if (collideWith(mainWindow.getNeko())) {
+        if (collideWith(box.getNeko())) {
             System.out.printf("고양이와 충돌: %.2f %.2f\n", x, y);
             show = false;
         }
