+++ src/net/honux/neko/Input.java
@@ -6,13 +6,13 @@ import java.awt.event.MouseMotionListener;
 
 public class Input implements MouseListener, MouseMotionListener {
 
-    private MainWindow m;
+    private Box box;
 
     private int mouseX;
     private int mouseY;
 
-    public Input(MainWindow m) {
-        this.m = m;
+    public Input(Box box) {
+        this.box = box;
     }
 
     public int getMouseX() {
@@ -25,7 +25,7 @@ public class Input implements MouseListener, MouseMotionListener {
 
     @Override
     public void mouseClicked(MouseEvent e) {
-        m.addCoin(e.getX(), e.getY());
+        box.addCoin(e.getX(), e.getY());
     }
 
     @Override
